+++ speech_to_text/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 7.0.0
+
+### New
+* support for speech recognition on MacOS
+
+### Fix
+* Remove support for deprecated v1 Android registration [#545](https://github.com/csdcorp/speech_to_text/issues/545)
+* Update deprecated Android build settings
+
 ## 6.6.2
 
 ### New